+++ aws-opensearchserverless/lib/collection.ts
@@ -1,0 +1,13 @@
+import {Construct} from 'constructs';
+import {ExtendedConstruct} from '../../aws-cdk';
+
+export interface CollectionProps {
+  readonly name: string;
+}
+
+export class Collection extends ExtendedConstruct {
+  constructor(scope: Construct, id: string, props: CollectionProps) {
+    super(scope, id);
+    console.log(props);
+  }
+}